requirements.txt
streamlit==1.32.2
Pillow==11.2.1
requests==2.31.0
beautifulsoup4==4.12.3
pandas==2.2.0
python-dotenv==1.0.1
tqdm==4.66.2
selenium==4.18.1
webdriver-manager==4.0.1 